requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.1
qrcode[pil]==7.4.2
pillow==10.4.0
numpy==2.1.1